lynx: add product page and autoscroll on first screen
This commit does two things:
 - It adds an autoscroll timer to the lynx page
 - It adds a product page for the lynx L1 and L2 volumes

The autoscroll timer scrolls the page on screen size down
if we are on the first screen (brand image) and the user does
not scroll within 1 second.
This improves the UX when a user opens the brand page the first time.
The user may not know that scrolling down will show all available
products. Therefore, the autoscroll points to the user directly to the
product if he/she does nothing.

Further, this commit adds a gallery slide-view product page.
On the left, there is a gallery slide-view of product images.
On the right, there is the product name, some key-properties
and an optional description of the product.